Move invitation mail to api
