+++ src/server/templates/invite.ejs
@@ -1,5 +1,1 @@
-<% include mailheader %>
-
-<%= inviter %> invited you to join URL "><%= user %>/<%= repo %> in ReviewNinja.
-
-<% include mailfooter %>
+<%= inviter %> invited you to join <a href="<%= url %>"><%= user %>/<%= repo %> in ReviewNinja.
