+++ src/tests/server/services/notification.js
@@ -16,7 +16,7 @@ var Settings = require('mongoose').model('Settings');
 var User = require('mongoose').model('User');
 var pullRequest = require('../../../server/services/pullRequest');
 
-describe('notification:', function() {
+describe('notification:sendmail', function() {
     it('should render the email content for new_issue', function(done) {
         var expectedcontent = '<h1>ReviewNinja</h1>\n\n\nA new issue has been raised for <a href="testurl">testuser/testrepo #1</a> by testsenderlogin.\n\n<p>--</p>\n<p>Automatic notification by ReviewNinja.<br>\nVisit <a href="http://www.review.ninja">review.ninja</a> to hear more about review.ninja.</p>\n\n';
 
@@ -100,61 +100,6 @@ describe('notification:', function() {
         settingsStub.restore();
         pullRequestStub.restore();
         userStub.restore();
-        nodemailerStub.restore();
-
-        done();
-    });
-
-    it('should send an invitation to the collaborator', function(done) {
-        var githubStub = sinon.stub(github, 'call', function(arg, done) {
-            assert.equal(arg.obj, 'user');
-            assert.equal(arg.fun, 'getFrom');
-            assert.equal(arg.arg.user, 'invitee');
-            assert.equal(arg.token, 'token');
-            done(null, { email: 'email@testmail.com'});
-        });
-
-        var transporter = {
-            sendMail: function(mailOptions, done) {
-                console.log('asdf', mailOptions);
-                // assert.equal();
-                assert.equal(mailOptions.to, 'email@testmail.com');
-                assert.equal(mailOptions.subject, 'You are invited to ReviewNinja');
-                assert.equal(mailOptions.html, '<h1>ReviewNinja</h1>\n\n\ninviter invited you to join URL ">' +
-                    '[object Object]/repo in ReviewNinja.\n\n<p>--</p>\n<p>Automatic notification by ReviewNinja.<br>' +
-                    '\nVisit <a href="http://www.review.ninja">review.ninja</a> to hear more about review.ninja.</p>\n\n');
-                done();
-            },
-            close: sinon.spy()
-        };
-        var nodemailerStub = sinon.stub(nodemailer, 'createTransport').returns(transporter);
-
-        notification.invite('user', 'repo', 'invitee', 'inviter', 'token');
-
-        assert(nodemailer.createTransport.called, 'transporter not created');
-        assert(transporter.close.called, 'mail not send via transporter');
-
-        githubStub.restore();
-        nodemailerStub.restore();
-
-        done();
-    });
-
-    it('should not send an invitation of there is no email of the collaborator', function(done) {
-        var githubStub = sinon.stub(github, 'call', function(arg, done) {
-            done(null, { email: ''});
-        });
-        var transporter = {
-            sendMail: sinon.stub()
-        };
-        var nodemailerStub = sinon.stub(nodemailer, 'createTransport').returns(transporter);
-
-        notification.invite(null, null, null, null, null);
-
-        assert(nodemailer.createTransport.notCalled, 'transporter called and created even if no email');
-        assert(transporter.sendMail.notCalled, 'mail send via transporter even if no email');
-
-        githubStub.restore();
         nodemailerStub.restore();
 
         done();
